requirements.txt

streamlit
python-docx
PyMuPDF
transformers
sentence-transformers
faiss-cpu
gTTS
reportlab
